rotating body: toggle-able options how
Facebutton 1 - Left motion controller:
toggles between 2 ways of controlling the rotation of the body.
Method 1: Via the z angle of the HMD
Method 2: Via the 2 hands. Forward direction of body coincides with the perpendicular vector to the vector connecting left and right hand (projected onto the XY plane, so we deal with 2D vectors which in this case results in the z angle)